refactor: validate 함수 service로 분리
